prd-agi-fuzzy/requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
scipy>=1.10.0
plotly>=5.14.0
pandas>=2.0.0
requests>=2.28.0
torch>=2.0.0
python-dotenv>=1.0.0
scikit-fuzzy>=0.4.2
